backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
boto3==1.34.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0

# ➕ Required for BERT-based semantic scoring
sentence-transformers==2.2.2
torch==2.1.0  # or compatible with your system's CUDA/CPU

# ➕ New: Keyword extraction from job descriptions
keybert==0.7.0

# ➕ For BERT scoring + keyword extraction
sentence-transformers==2.2.2
torch==2.1.0         # or compatible with your system
keybert==0.7.0
reportlab==3.6.2